consolidate functions and add checks to more inputs
