+++ dicom2nifti/common.py
@@ -104,7 +104,7 @@ def is_multiframe_dicom(dicom_input):
     :param dicom_input: directory with dicom files for 1 scan
     """
     # read dicom header
-    header = dicom_input[0]
+    header = dicom_input[0][0] if isinstance(dicom_input[0], list) else dicom_input[0]
 
     if Tag(0x0002, 0x0002) not in header.file_meta:
         return False
@@ -774,3 +774,119 @@ def get_nifti_data(nifti_image):
     Function that replicates the deprecated nifti.get_data behavior
     """
     return numpy.asanyarray(nifti_image.dataobj)
+
+
+def remove_duplicate_slices(dicoms):
+    """
+    Search dicoms for localizers and delete them
+    """
+    # Loop overall files and build dict
+
+    dicoms_dict = {}
+    filtered_dicoms = []
+    # in case of multiframe this check cannot be done
+    if 'ImageOrientationPatient' not in dicoms[0]:
+        return dicoms
+
+    for dicom_ in dicoms:
+        if tuple(dicom_.ImagePositionPatient) not in dicoms_dict:
+            dicoms_dict[tuple(dicom_.ImagePositionPatient)] = dicom_
+            filtered_dicoms.append(dicom_)
+        else:
+            if numpy.array_equal(dicom_.pixel_array,
+                                 dicoms_dict[tuple(dicom_.ImagePositionPatient)].pixel_array):
+                logger.warning('Removing duplicate slice from series')
+            else:
+                filtered_dicoms.append(dicom_)
+    return filtered_dicoms
+
+
+def remove_localizers_by_imagetype(dicoms):
+    """
+    Search dicoms for localizers and delete them
+    """
+    # Loop overall files and build dict
+    filtered_dicoms = []
+    for dicom_ in dicoms:
+        if 'ImageType' in dicom_ and 'LOCALIZER' in dicom_.ImageType:
+            continue
+        # 'Projection Image' are Localizers for CT only see MSMET-234
+        if 'CT' in dicom_.Modality and 'ImageType' in dicom_ and 'PROJECTION IMAGE' in dicom_.ImageType:
+            continue
+        filtered_dicoms.append(dicom_)
+    return filtered_dicoms
+
+
+def remove_localizers_by_orientation(dicoms):
+    """
+    Removing localizers based on the orientation.
+    This is needed as in some cases with ct data there are some localizer/projection type images that cannot
+    be distiguished by the dicom headers. This is why we kick out all orientations that do not have more than 4 files
+    4 is the limit anyway for converting to nifti on our case
+    """
+    orientations = []
+    sorted_dicoms = {}
+
+    # in case of multiframe this check cannot be done
+    if 'ImageOrientationPatient' not in dicoms[0]:
+        return dicoms
+
+    # Loop overall files and build dict
+    for dicom_header in dicoms:
+        # Create affine matrix (http://nipy.sourceforge.net/nibabel/dicom/dicom_orientation.html#dicom-slice-affine)
+        image_orient1 = numpy.array(dicom_header.ImageOrientationPatient)[0:3]
+        image_orient2 = numpy.array(dicom_header.ImageOrientationPatient)[3:6]
+        image_orient_combined = (image_orient1.tolist(), image_orient2.tolist())
+        found_orientation = False
+        for orientation in orientations:
+            if numpy.allclose(image_orient_combined[0], numpy.array(orientation[0]), rtol=0.001, atol=0.001) \
+                    and numpy.allclose(image_orient_combined[1], numpy.array(orientation[1]), rtol=0.001,
+                                       atol=0.001):
+                sorted_dicoms[str(orientation)].append(dicom_header)
+                found_orientation = True
+                break
+        if not found_orientation:
+            orientations.append(image_orient_combined)
+            sorted_dicoms[str(image_orient_combined)] = [dicom_header]
+
+    # if there are multiple possible orientations delete orientations where there are less than 4 files
+    # we don't convert anything less that that anyway
+
+    if len(sorted_dicoms) > 1:
+        filtered_dicoms = []
+        for orientation in sorted_dicoms.keys():
+            if len(sorted_dicoms[orientation]) >= 4:
+                filtered_dicoms.extend(sorted_dicoms[orientation])
+        return filtered_dicoms
+    else:
+        return next(iter(sorted_dicoms.values()))
+
+
+def initial_dicom_checks(dicom_input, mosaic=False):
+    if len(dicom_input) <= 0:
+        raise ConversionError('NO_DICOM_FILES_FOUND')
+
+    # remove duplicate slices based on position and data
+    dicom_input = remove_duplicate_slices(dicom_input)
+    # remove localizers based on image type
+    dicom_input = remove_localizers_by_imagetype(dicom_input)
+    # if no dicoms remain we should raise exception
+    if len(dicom_input) < 1:
+        raise ConversionValidationError('TOO_FEW_SLICES/LOCALIZER')
+
+    if settings.validate_slicecount and not mosaic:
+        validate_slicecount(dicom_input)
+        # remove_localizers based on image orientation (only valid if slicecount is validated)
+        dicom_input = remove_localizers_by_orientation(dicom_input)
+        # validate all the dicom files for correct orientations
+        validate_slicecount(dicom_input)
+
+    if settings.validate_orientation:
+        # validate that all slices have the same orientation
+        validate_orientation(dicom_input)
+
+    if settings.validate_orthogonal:
+        # validate that we have an orthogonal image (to detect gantry tilting etc)
+        validate_orthogonal(dicom_input)
+
+    return dicom_input
