+++ dicom2nifti/convert_generic.py
@@ -10,17 +10,16 @@ import numpy
 
 from pydicom.tag import Tag
 
-from .common import validate_slicecount, validate_orthogonal,  validate_orientation, \
-    validate_slice_increment, validate_instance_number, is_slice_increment_inconsistent, \
-    sort_dicoms, get_volume_pixeldata, create_affine, set_tr_te, is_multiframe_dicom
+from .common import is_slice_increment_inconsistent, get_volume_pixeldata, \
+    create_affine, set_tr_te, is_multiframe_dicom, validate_slice_increment, \
+    validate_instance_number, initial_dicom_checks, sort_dicoms
 from .settings import Dicom2NiftiSettings as settings
 from .resample import resample_nifti_images
-from .exceptions import ConversionError, ConversionValidationError
 
 logger = logging.getLogger(__name__)
 
 
-def generic_dicom_to_nifti(dicom_input, output_file):
+def generic_dicom_to_nifti(dicom_input, output_file, run_checks=True):
     """
     This function will convert single frame dicom series to a nifti
 
@@ -28,42 +27,19 @@ def generic_dicom_to_nifti(dicom_input, output_file):
 
     :param output_file: filepath to the output nifti
     :param dicom_input: directory with the dicom files for a single scan, or list of read in dicoms
+    :param run_checks:
     """
-    if len(dicom_input) <= 0:
-        raise ConversionError('NO_DICOM_FILES_FOUND')
-
-    # remove duplicate slices based on position and data
-    dicom_input = remove_duplicate_slices(dicom_input)
-    # remove localizers based on image type
-    dicom_input = remove_localizers_by_imagetype(dicom_input)
-    # if no dicoms remain we should raise exception
-    if len(dicom_input) < 1:
-        raise ConversionValidationError('TOO_FEW_SLICES/LOCALIZER')
-
-    if settings.validate_slicecount:
-        validate_slicecount(dicom_input)
-        # remove_localizers based on image orientation (only valid if slicecount is validated)
-        dicom_input = remove_localizers_by_orientation(dicom_input)
-        # validate all the dicom files for correct orientations
-        validate_slicecount(dicom_input)
-    if settings.validate_orientation:
-        # validate that all slices have the same orientation
-        validate_orientation(dicom_input)
-    if settings.validate_orthogonal:
-        # validate that we have an orthogonal image (to detect gantry tilting etc)
-        validate_orthogonal(dicom_input)
+    if run_checks:
+        dicom_input = initial_dicom_checks(dicom_input)
 
     if not is_multiframe_dicom(dicom_input):
         # sort the dicoms
         dicom_input = sort_dicoms(dicom_input)
 
     # validate slice increment inconsistent
-    slice_increment_inconsistent = False
     if settings.validate_slice_increment:
         # validate that all slices have a consistent slice increment
         validate_slice_increment(dicom_input)
-    elif is_slice_increment_inconsistent(dicom_input):
-        slice_increment_inconsistent = True
 
     if settings.validate_instance_number:
         # validate that all slices have a consistent instance_number
@@ -71,9 +47,9 @@ def generic_dicom_to_nifti(dicom_input, output_file):
 
     # if inconsistent increment and we allow resampling then do the resampling based conversion
     # to maintain the correct geometric shape
-    if slice_increment_inconsistent and settings.resample:
+    if is_slice_increment_inconsistent(dicom_input) and settings.resample:
         if is_multiframe_dicom(dicom_input):
-            raise Exception('Conversion of inconsistent slice increment with resampling not supported for mutlframe')
+            raise Exception('Conversion of inconsistent slice increment with resampling not supported for multiframe')
             # TODO add support for this if it actually exists
         nii_image, max_slice_increment = _convert_slice_increment_inconsistencies(dicom_input)
     # do the normal conversion
@@ -117,92 +93,6 @@ def create_rgba_nifti(rgb_data, affine):
     # copy used to force fresh internal structure
     rgba_data = rgba_data.copy().view(dtype=rgba_dtype).reshape(shape_3d)
     return nibabel.Nifti1Image(rgba_data, affine)
-
-
-def remove_duplicate_slices(dicoms):
-    """
-    Search dicoms for localizers and delete them
-    """
-    # Loop overall files and build dict
-
-    dicoms_dict = {}
-    filtered_dicoms = []
-    # in case of multiframe this check cannot be done
-    if 'ImageOrientationPatient' not in dicoms[0]:
-        return dicoms
-
-    for dicom_ in dicoms:
-        if tuple(dicom_.ImagePositionPatient) not in dicoms_dict:
-            dicoms_dict[tuple(dicom_.ImagePositionPatient)] = dicom_
-            filtered_dicoms.append(dicom_)
-        else:
-            if numpy.array_equal(dicom_.pixel_array,
-                                 dicoms_dict[tuple(dicom_.ImagePositionPatient)].pixel_array):
-                logger.warning('Removing duplicate slice from series')
-            else:
-                filtered_dicoms.append(dicom_)
-    return filtered_dicoms
-
-
-def remove_localizers_by_imagetype(dicoms):
-    """
-    Search dicoms for localizers and delete them
-    """
-    # Loop overall files and build dict
-    filtered_dicoms = []
-    for dicom_ in dicoms:
-        if 'ImageType' in dicom_ and 'LOCALIZER' in dicom_.ImageType:
-            continue
-        # 'Projection Image' are Localizers for CT only see MSMET-234
-        if 'CT' in dicom_.Modality and 'ImageType' in dicom_ and 'PROJECTION IMAGE' in dicom_.ImageType:
-            continue
-        filtered_dicoms.append(dicom_)
-    return filtered_dicoms
-
-
-def remove_localizers_by_orientation(dicoms):
-    """
-    Removing localizers based on the orientation.
-    This is needed as in some cases with ct data there are some localizer/projection type images that cannot
-    be distiguished by the dicom headers. This is why we kick out all orientations that do not have more than 4 files
-    4 is the limit anyway for converting to nifti on our case
-    """
-    orientations = []
-    sorted_dicoms = {}
-
-    # in case of multiframe this check cannot be done
-    if 'ImageOrientationPatient' not in dicoms[0]:
-        return dicoms
-
-    # Loop overall files and build dict
-    for dicom_header in dicoms:
-        # Create affine matrix (http://nipy.sourceforge.net/nibabel/dicom/dicom_orientation.html#dicom-slice-affine)
-        image_orient1 = numpy.array(dicom_header.ImageOrientationPatient)[0:3]
-        image_orient2 = numpy.array(dicom_header.ImageOrientationPatient)[3:6]
-        image_orient_combined = (image_orient1.tolist(), image_orient2.tolist())
-        found_orientation = False
-        for orientation in orientations:
-            if numpy.allclose(image_orient_combined[0], numpy.array(orientation[0]), rtol=0.001, atol=0.001) \
-                    and numpy.allclose(image_orient_combined[1], numpy.array(orientation[1]), rtol=0.001,
-                                       atol=0.001):
-                sorted_dicoms[str(orientation)].append(dicom_header)
-                found_orientation = True
-                break
-        if not found_orientation:
-            orientations.append(image_orient_combined)
-            sorted_dicoms[str(image_orient_combined)] = [dicom_header]
-
-    # if there are multiple possible orientations delete orientations where there are less than 4 files
-    # we don't convert anything less that that anyway
-
-    if len(sorted_dicoms) > 1:
-        filtered_dicoms = []
-        for orientation in sorted_dicoms.keys():
-            if len(sorted_dicoms[orientation]) >= 4:
-                filtered_dicoms.extend(sorted_dicoms[orientation])
-        return filtered_dicoms
-    else:
-        return next(iter(sorted_dicoms.values()))
 
 
 def _convert_slice_increment_inconsistencies(dicom_input):
